changed default category route to /category/outpatient
